ci: run all tests
Previously, tests actually using the Landlock kernel API were allowed to
fail if Landlock was not supported in the test environment. This was the
case on Ubuntu 18.04 runners in the GitHub Actions environment, because
of the old Linux kernel version used.

Hence, actual Landlock functionality was not tested (in CI, at least).

The Ubuntu 22.04 runners, which come with a new kernel version which
does have support for Landlock, also failed to execute the tests (or
rather, executed them with expected failures). It turns out the version
of Moby/Docker used in this environment comes with a `seccomp` profile
which doesn't allow the Landlock syscalls to be performed (this was
added in Moby version 22.06-beta.0).

However, on GitHub Actions, we can configure how `docker build` is
invoked. This allows the `seccomp` profile to be configured as
`unconfined`, effectively allowing the use of Landlock APIs (and more,
of course).

This patch

- Configures the `Haskell-CI` GitHub Actions workflow to run in the
  `ubuntu-22.04` environment
- Configures GitHub Actions to pass the `--security-opt
  seccomp=unconfined` argument to `docker build`
- Removes the "expected failure" wrappers of all tests

See: https://github.com/moby/moby/blob/v20.10.17/profiles/seccomp/default.json
See: https://github.com/moby/moby/commit/af819bf623a8d414289b24f9817c6317cf3f96d5