tests: check typeclass laws of instances using `quickcheck-classes-base`
Instead of having some hand-rolled minimal laws check for `Storable`,
use `quickcheck-classes-base` to validate a bunch of laws on a bunch of
datatypes.

Note: the `quickcheck-classes-base` package is used instead of
`quickcheck-classes` because the latter take (much) longer to build in
CI (because of its dependency on `aeson`). We only need checks for
`base` classes anyway.

See: https://hackage.haskell.org/package/quickcheck-classes-base