chore: move GHC `-Wall` option into `cabal.project`
Since this is a flag that's only used for development purposed and
shouldn't impact builds of the packages when used as a dependency,
moving it from the packages' Cabal file into the project's
configuration.

Not (yet) done for C compiler invocations, since unclear how to achieve
this.

See: https://mail.haskell.org/pipermail/haskell-cafe/2022-August/135476.html