landlock: remove superfluous constraints
Since `restrictSelfFlagToBit` and `addRuleFlagToBit` can't ever be
called (there's no value for their argument type), their result type is
(unconstrained) `a`. Whilst it should get a `Num a` constraint once
there would be some flags defined and the functions extended to handle
them, this is currently not the case.

Cfr. `accessFsFlagToBit` and `createRulesetFlagToBit`.