landlock: import `linux/landlock.h` from Linux v5.18.19
Instead of relying on the Landlock UAPI header(s) found on a build
system, include them in the tree (and use them later).

The rationale is that this library should support multiple versions of
the Landlock ABI over time, but when building with a system-provided
`landlock.h` header, we can't make assumptions about definitions found
in it.

Using conditional compilation, based on whatever is found in the
system's kernel headers, would result in this library exposing a
different API depending on where/when it's built, which is...
sub-optimal.

The Linux kernel syscall APIs are (supposed to be) backwards compatible,
so if we ship a version of `landlock` with a header from a version of
Linux that's newer than the kernel version on which the code will run,
this should not be an issue.

Retrieved from Linux Git, tag `v5.18.19`, revision
22a992953741ad79c07890d3f4104585e52ef26b, file
`include/uapi/linux/landlock.h`.

This file is licensed under the GPL-2.0 with Linux-syscall-note. The
relevant license files are included as well. Copyright of the file
belongs to the original authors.

See: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/tree/include/uapi/linux/landlock.h?h=v5.18.19&id=22a992953741ad79c07890d3f4104585e52ef26b