landlock: add `landlocked` flag
Setting the flag to `False` will result in the `landlocked` utility not
to be built.

This could be useful if at some point one wants to build the library in
an environment where the `landlocked` `Build-Depends` would conflict
with other dependencies. Or if one simply doesn't want the utility to be
installed.