landlock: rework `accessFsFlags` implementation
Instead of relying on the ordering of constructors in the datatype,
maintain a mapping of flag to ABI version in the code, and construct the
mappings from ABI version to flags set accordingly.

Also, add some unit-tests to check this is done correctly. Right now
soem of these rely on the datatype constructors ordering, hence work
like the original implementation, but that's fine.

(cherry picked from commit 2f09423f00955d8f63cf420ed381a970ab496390)