landlock: prefix syscall bindings
The syscall bindings in `landlock/cbits/hs-landlock.c` were using the
syscall names as-is, which could conflict with implementations in, e.g.,
`libc`.

Furthermore, the binding to `landlock_restrict_self` is using `libpsx`'
interface, hence it's not even a plain binding to the
`landlock_restrict_self` syscall.

So, this patch changes all bindings to use the `hs_` prefix (as was
already the case for `hs_landlock_prctl`). Internally, they'll use an
existing binding (if defined), otherwise use `syscall(2)` as was the
original code.