psx: work-around `hsc2hs` linking issue
On some platforms, building a `hsc2hs` program requires `libgcc`, and on
some platforms, `libgcc` uses `sigfillset`. For some reason (a bug, I'd
say), `cabal` passes `ld-options` to the compiler when building the
generated C file. In our case, this causes `sigfillset` to be wrapped
(which is desired for the actual `psx` library), but since
`__wrap_sigfillset` is never defined, the build fails.

This work-around tells `hsc2hs` to include a small C file in its
generated sources which defines `__wrap_sigfillset` as a proxy for
`__real_sigfillset`.

This requires the Cabal format to be bumped to 3.6.

See: https://github.com/haskell/cabal/issues/8824