landlock: properly zero-out structs after allocation
When allocating a struct of a specific size, but only filling in some
members, everything should be zeroed out first (before, e.g., passing
the struct to the kernel).

This was not an issue until now, since the relevant API structures are
always fully filled, however, with an upcoming Landlock struct change,
this is required: when not zeroing the memory, some random values would
be passed to the kernel, resulting in an error (this could only be
reproduced on GHC 8.10 and 9.0, though, it looks like more recent
versions have `alloca` return zeroed memory).

Where before `Foreign.Marshal.Utils.with` was used to get a pointer to a
structure, `with0` is introduced which basically does the same
(`alloca`, `poke`, run the callback), but uses `fillBytes` to zero out
the `alloca`ted memory before `poke`ing the `Storable` in place.